+++ modules/core/CMakeLists.txt
@@ -3,6 +3,10 @@ set(the_description "The Core Functionality")
 ocv_add_dispatched_file(mathfuncs_core SSE2 AVX AVX2)
 ocv_add_dispatched_file(stat SSE4_2 AVX2)
 
+# dispatching for accuracy tests
+ocv_add_dispatched_file_force_all(test_intrin128 TEST SSE2 SSE3 SSSE3 SSE4_1 SSE4_2 AVX FP16 AVX2)
+ocv_add_dispatched_file_force_all(test_intrin256 TEST AVX2)
+
 ocv_add_module(core
                OPTIONAL opencv_cudev
                WRAP java python js)
