+++ modules/core/test/test_intrin.cpp
@@ -2,101 +2,100 @@
 // It is subject to the license terms in the LICENSE file found in the top-level directory
 // of this distribution and at http://opencv.org/license.html.
 #include "test_precomp.hpp"
-#include "test_intrin.simd.hpp"
 
-#define CV_CPU_SIMD_FILENAME "test_intrin.simd.hpp"
-#define CV_CPU_DISPATCH_MODE FP16
-#include "opencv2/core/private/cv_cpu_include_simd_declarations.hpp"
+#include "test_intrin128.simd.hpp"
+#include "test_intrin128.simd_declarations.hpp"
 
-#define CV_CPU_DISPATCH_MODE AVX2
-#include "opencv2/core/private/cv_cpu_include_simd_declarations.hpp"
+#undef CV_CPU_DISPATCH_MODES_ALL
+
+#include "opencv2/core/cv_cpu_dispatch.h"
+#include "test_intrin256.simd.hpp"
+#include "test_intrin256.simd_declarations.hpp"
+
 
 namespace opencv_test { namespace hal {
-using namespace CV_CPU_OPTIMIZATION_NAMESPACE;
 
-TEST(hal_intrin, uint8x16)
-{ test_hal_intrin_uint8(); }
+#define CV_CPU_CALL_BASELINE_(fn, args)  CV_CPU_CALL_BASELINE(fn, args)
 
-TEST(hal_intrin, int8x16)
-{ test_hal_intrin_int8(); }
+#define DISPATCH_SIMD128(fn, cpu_opt) do { \
+    CV_CPU_CALL_ ## cpu_opt ## _(fn, ()); \
+    throw SkipTestException("SIMD128 (" #cpu_opt ") is not available or disabled"); \
+} while(0)
 
-TEST(hal_intrin, uint16x8)
-{ test_hal_intrin_uint16(); }
+#define DISPATCH_SIMD256(fn, cpu_opt) do { \
+    CV_CPU_CALL_ ## cpu_opt ## _(fn, ()); \
+    throw SkipTestException("SIMD256 (" #cpu_opt ") is not available or disabled"); \
+} while(0)
 
-TEST(hal_intrin, int16x8)
-{ test_hal_intrin_int16(); }
+#define DEFINE_SIMD_TESTS(simd_size, cpu_opt) \
+TEST(hal_intrin ## simd_size, uint8x16_ ## cpu_opt)  { DISPATCH_SIMD ## simd_size(test_hal_intrin_uint8, cpu_opt); } \
+TEST(hal_intrin ## simd_size, int8x16_ ## cpu_opt)   { DISPATCH_SIMD ## simd_size(test_hal_intrin_int8, cpu_opt); } \
+TEST(hal_intrin ## simd_size, uint16x8_ ## cpu_opt)  { DISPATCH_SIMD ## simd_size(test_hal_intrin_uint16, cpu_opt); } \
+TEST(hal_intrin ## simd_size, int16x8_ ## cpu_opt)   { DISPATCH_SIMD ## simd_size(test_hal_intrin_int16, cpu_opt); } \
+TEST(hal_intrin ## simd_size, int32x4_ ## cpu_opt)   { DISPATCH_SIMD ## simd_size(test_hal_intrin_int32, cpu_opt); } \
+TEST(hal_intrin ## simd_size, uint32x4_ ## cpu_opt)  { DISPATCH_SIMD ## simd_size(test_hal_intrin_uint32, cpu_opt); } \
+TEST(hal_intrin ## simd_size, uint64x2_ ## cpu_opt)  { DISPATCH_SIMD ## simd_size(test_hal_intrin_uint64, cpu_opt); } \
+TEST(hal_intrin ## simd_size, int64x2_ ## cpu_opt)   { DISPATCH_SIMD ## simd_size(test_hal_intrin_int64, cpu_opt); } \
+TEST(hal_intrin ## simd_size, float32x4_ ## cpu_opt) { DISPATCH_SIMD ## simd_size(test_hal_intrin_float32, cpu_opt); } \
+TEST(hal_intrin ## simd_size, float64x2_ ## cpu_opt) { DISPATCH_SIMD ## simd_size(test_hal_intrin_float64, cpu_opt); } \
 
-TEST(hal_intrin, int32x4)
-{ test_hal_intrin_int32(); }
+namespace intrin128 {
 
-TEST(hal_intrin, uint32x4)
-{ test_hal_intrin_uint32(); }
+DEFINE_SIMD_TESTS(128, BASELINE)
 
-TEST(hal_intrin, uint64x2)
-{ test_hal_intrin_uint64(); }
+#if defined CV_CPU_DISPATCH_COMPILE_SSE2 || defined CV_CPU_BASELINE_COMPILE_SSE2
+DEFINE_SIMD_TESTS(128, SSE2)
+#endif
+#if defined CV_CPU_DISPATCH_COMPILE_SSE3 || defined CV_CPU_BASELINE_COMPILE_SSE3
+DEFINE_SIMD_TESTS(128, SSE3)
+#endif
+#if defined CV_CPU_DISPATCH_COMPILE_SSSE3 || defined CV_CPU_BASELINE_COMPILE_SSSE3
+DEFINE_SIMD_TESTS(128, SSSE3)
+#endif
+#if defined CV_CPU_DISPATCH_COMPILE_SSE4_1 || defined CV_CPU_BASELINE_COMPILE_SSE4_1
+DEFINE_SIMD_TESTS(128, SSE4_1)
+#endif
+#if defined CV_CPU_DISPATCH_COMPILE_SSE4_2 || defined CV_CPU_BASELINE_COMPILE_SSE4_2
+DEFINE_SIMD_TESTS(128, SSE4_2)
+#endif
+#if defined CV_CPU_DISPATCH_COMPILE_AVX || defined CV_CPU_BASELINE_COMPILE_AVX
+DEFINE_SIMD_TESTS(128, AVX)
+#endif
+#if defined CV_CPU_DISPATCH_COMPILE_AVX2 || defined CV_CPU_BASELINE_COMPILE_AVX2
+DEFINE_SIMD_TESTS(128, AVX2)
+#endif
 
-TEST(hal_intrin, int64x2)
-{ test_hal_intrin_int64(); }
-
-TEST(hal_intrin, float32x4)
-{ test_hal_intrin_float32(); }
-
-TEST(hal_intrin, float64x2)
-{ test_hal_intrin_float64(); }
-
-TEST(hal_intrin, float16x8)
+TEST(hal_intrin128, float16x8_FP16)
 {
     CV_CPU_CALL_FP16_(test_hal_intrin_float16, ());
     throw SkipTestException("Unsupported hardware: FP16 is not available");
 }
 
-#define DISPATCH_SIMD_MODES AVX2
-#define DISPATCH_SIMD_NAME "SIMD256"
-#define DISPATCH_SIMD(fun)                              \
-    do {                                                \
-        CV_CPU_DISPATCH(fun, (), DISPATCH_SIMD_MODES);  \
-        throw SkipTestException(                        \
-            "Unsupported hardware: "                    \
-            DISPATCH_SIMD_NAME                          \
-            " is not available"                         \
-        );                                              \
-    } while(0)
+} // namespace intrin128
 
-TEST(hal_intrin256, uint8x32)
-{ DISPATCH_SIMD(test_hal_intrin_uint8); }
 
-TEST(hal_intrin256, int8x32)
-{ DISPATCH_SIMD(test_hal_intrin_int8); }
+namespace intrin256 {
 
-TEST(hal_intrin256, uint16x16)
-{ DISPATCH_SIMD(test_hal_intrin_uint16); }
 
-TEST(hal_intrin256, int16x16)
-{ DISPATCH_SIMD(test_hal_intrin_int16); }
+// Not available due missing C++ backend for SIMD256
+//DEFINE_SIMD_TESTS(256, BASELINE)
 
-TEST(hal_intrin256, uint32x8)
-{ DISPATCH_SIMD(test_hal_intrin_uint32); }
+//#if defined CV_CPU_DISPATCH_COMPILE_AVX
+//DEFINE_SIMD_TESTS(256, AVX)
+//#endif
 
-TEST(hal_intrin256, int32x8)
-{ DISPATCH_SIMD(test_hal_intrin_int32); }
+#if defined CV_CPU_DISPATCH_COMPILE_AVX2 || defined CV_CPU_BASELINE_COMPILE_AVX2
+DEFINE_SIMD_TESTS(256, AVX2)
+#endif
 
-TEST(hal_intrin256, uint64x4)
-{ DISPATCH_SIMD(test_hal_intrin_uint64); }
-
-TEST(hal_intrin256, int64x4)
-{ DISPATCH_SIMD(test_hal_intrin_int64); }
-
-TEST(hal_intrin256, float32x8)
-{ DISPATCH_SIMD(test_hal_intrin_float32); }
-
-TEST(hal_intrin256, float64x4)
-{ DISPATCH_SIMD(test_hal_intrin_float64); }
-
-TEST(hal_intrin256, float16x16)
+TEST(hal_intrin256, float16x16_FP16)
 {
-    if (!CV_CPU_HAS_SUPPORT_FP16)
-        throw SkipTestException("Unsupported hardware: FP16 is not available");
-    DISPATCH_SIMD(test_hal_intrin_float16);
+    //CV_CPU_CALL_FP16_(test_hal_intrin_float16, ());
+    CV_CPU_CALL_AVX2_(test_hal_intrin_float16, ());
+    throw SkipTestException("Unsupported hardware: FP16 is not available");
 }
+
+
+} // namespace intrin256
 
 }} // namespace
