+++ modules/core/test/test_precomp.hpp
@@ -11,6 +11,5 @@
 #include "opencv2/core/cvdef.h"
 #include "opencv2/core/private.hpp"
 #include "opencv2/core/hal/hal.hpp"
-#include "opencv2/core/hal/intrin.hpp"
 
 #endif
